chore: merge debug.js + refactor
